+++ sandpack-react/src/_docs-index/components.ts
@@ -1,0 +1,4 @@
+export * from "../components";
+export * from "../common";
+export * from "../presets";
+export * from "../types";