+++ sandpack-react/src/components/index.ts
@@ -1,0 +1,6 @@
+export * from "./CodeEditor";
+export * from "./CodeViewer";
+export * from "./FileTabs";
+export * from "./Navigator";
+export * from "./Preview";
+export * from "./TranspiledCode";