+++ sandpack-react/src/types.ts
@@ -99,6 +99,7 @@ export type SandpackPredefinedTemplate =
   | "react"
   | "react-ts"
   | "vanilla"
+  | "vanilla-ts"
   | "vue"
   | "vue3";
 
@@ -170,6 +171,9 @@ export type SandpackThemeProp =
   | SandpackPartialTheme
   | "auto";
 
+/**
+ * @hidden
+ */
 export type DeepPartial<Type> = {
   [Property in keyof Type]?: DeepPartial<Type[Property]>;
 };