+++ website/docs/docusaurus.config.js
@@ -1,3 +1,8 @@
+const typeDocsDefault = {
+  readme: "none",
+  plugin: ["typedoc-plugin-no-inherit"],
+};
+
 // With JSDoc @type annotations, IDEs can provide config autocompletion
 /** @type {import('@docusaurus/types').DocusaurusConfig} */
 (
@@ -15,7 +20,95 @@
     themes: ["@codesandbox/sandpack-docusaurus"],
 
     plugins: [
-      'docusaurus-plugin-sass',
+      /**
+       * API Reference
+       */
+      [
+        "docusaurus-plugin-typedoc",
+        {
+          ...typeDocsDefault,
+          id: "components",
+          entryPoints: ["../../sandpack-react/src/_docs-index/components.ts"],
+          tsconfig: "../../sandpack-react/tsconfig.json",
+          out: "api/react/components",
+          indexTitle: "Components",
+          sidebar: {
+            categoryLabel: "Components",
+            fullNames: false,
+            readmeLabel: "Components",
+            position: 0,
+          },
+        },
+      ],
+      [
+        "docusaurus-plugin-typedoc",
+        {
+          ...typeDocsDefault,
+          id: "provider",
+          entryPoints: ["../../sandpack-react/src/_docs-index/provider.ts"],
+          tsconfig: "../../sandpack-react/tsconfig.json",
+          out: "api/react/provider",
+          indexTitle: "Provider",
+          sidebar: {
+            categoryLabel: "Provider",
+            fullNames: false,
+            readmeLabel: "Provider",
+            position: 1,
+          },
+        },
+      ],
+      [
+        "docusaurus-plugin-typedoc",
+        {
+          ...typeDocsDefault,
+          id: "hooks",
+          entryPoints: ["../../sandpack-react/src/_docs-index/hooks.ts"],
+          tsconfig: "../../sandpack-react/tsconfig.json",
+          out: "api/react/hooks",
+          indexTitle: "Hooks",
+          sidebar: {
+            categoryLabel: "Hooks",
+            fullNames: false,
+            readmeLabel: "Hooks",
+            position: 2,
+          },
+        },
+      ],
+      [
+        "docusaurus-plugin-typedoc",
+        {
+          ...typeDocsDefault,
+          id: "Theme",
+          entryPoints: ["../../sandpack-react/src/_docs-index/theme.ts"],
+          tsconfig: "../../sandpack-react/tsconfig.json",
+          out: "api/react/theme",
+          indexTitle: "Theme",
+          sidebar: {
+            categoryLabel: "Theme",
+            readmeLabel: "Theme",
+            position: 3,
+          },
+        },
+      ],
+      [
+        "docusaurus-plugin-typedoc",
+        {
+          ...typeDocsDefault,
+          id: "client",
+          entryPoints: ["../../sandpack-client/src/index.ts"],
+          tsconfig: "../../sandpack-client/tsconfig.json",
+          out: "api/client",
+          indexTitle: "Sandpack-client",
+          sidebar: {
+            categoryLabel: "Sandpack-client",
+            readmeLabel: "Sandpack-client",
+            position: 4,
+          },
+        },
+      ],
+      /** END */
+
+      "docusaurus-plugin-sass",
       [
         "@docusaurus/plugin-sitemap",
         {
@@ -47,7 +140,7 @@
               "https://github.com/codesandbox/sandpack/edit/main/website/docs",
           },
           theme: {
-            customCss: [require.resolve('./src/scss/custom.scss')],
+            customCss: [require.resolve("./src/scss/custom.scss")],
           },
         }),
       ],
@@ -56,9 +149,10 @@
     themeConfig:
       /** @type {import('@docusaurus/preset-classic').ThemeConfig} */
       ({
-        // sandpack: { theme: require("./src/css/sandpack") },
+        sandpack: { theme: require("./src/scss/sandpack-theme") },
         colorMode: {
           defaultMode: "dark",
+          disableSwitch: true,
         },
         navbar: {
           title: "Sandpack",
@@ -68,7 +162,7 @@
             srcDark: "img/logo.svg",
           },
           items: [
-            { to: "/releases", label: "Releases", position: "left" },
+            { to: "/releases", label: "Releases", position: "right" },
             {
               href: "https://github.com/codesandbox/sandpack",
               label: "GitHub",